Avoid referencing ns_msghdr_t::flags
This field member was wrongly named - it should have been msg_flags.
I'm planning to correct it, which is a compatibility break.

As this doesn't need to actually set any flags for sendmsg(), use
designated initialiser to get it automatically zero-filled, so this code
will work with both old and new naming.

For recvmsg() the flags are output-only, so no need to set it when
making the call.