Updated coap service to be compatible with mbed TLS 3.0 (#135)
Disabled key material export for now, needs to be updated later to new
key export function if needed. Made the SSL state to refer to private
state for now. Do not include config, instead include version.h that
is present in both 2.0 and 3.0 and will include config and defines version
macros.